Move `Analytics` data structs to `Core`
