+++ Sources/CLI/PRCommand.swift
@@ -121,7 +121,8 @@ class PRCommand: Command {
 				documentName: "pull-requests",
 				body: PullRequestAnalyticsData(
 					outputAnalyticsData: badondefileOutput.analyticsData,
-					startDate: startDatePointer.pointee
+					startDate: startDatePointer.pointee,
+					version: CommandLineTool.Constant.version
 				)
 			)
 		}
